Let libc++abi compile with gcc.
There was a single problem in cxa_demangle.cpp, where gcc would complain
`error: changes meaning of 'String'` about the line `typedef String String;`.
According to 3.3.7p2, this diagnostic is allowed (but not required, so clang
does not have to report this).

As a fix, make string_pair a template and pass String as template parameter.
This fixes the error with gcc and also removes some repetition from the code. 

No behavior change.


git-svn-id: svn://svn.chromium.org/llvm-project/libcxxabi/trunk@209909 0b72dbe1-c17f-4bc7-b9db-2b4152be0356