unwind: always export unw_local_addr_space
It seems that the remote unwinder is entirely unused at this moment.
unw_local_addr_space was referencing sThisAddressSpace which use to be a static
in global namespace.  It has since then become a member variable of
LocalAddressSpace.  Update this definition and always export it (needed to
implement unw_get_proc_info_by_ip for ARM).

git-svn-id: svn://svn.chromium.org/llvm-project/libcxxabi/trunk@229133 0b72dbe1-c17f-4bc7-b9db-2b4152be0356