Add .eh_frame_hdr search to Linux unwinder.
This improves the performance of unwinding on DWARF based targets. The
32-bit x86 support for scanning the full eh_frame
(CFI_Parser::findFDE) apparently does not work (at least not on
Linux). Since the eh_frame_hdr code delegates to that, this still
doesn't work for x86 Linux, but it has been tested on x86_64 Linux and
aarch64 Android.

git-svn-id: svn://svn.chromium.org/llvm-project/libcxxabi/trunk@230802 0b72dbe1-c17f-4bc7-b9db-2b4152be0356