EH: fix register usage for SjLj
When using SjLj EH, do not use __builtin_eh_return_regno, map directly to the
ID.  This would work on some targets, particularly those where the non-SjLj EH
personality used the same register mapping (0 -> 0, 1 -> 1).  However, this is
not guaranteed.  Avoiding the use of the builtin enables the use of libc++ with
SjLj EH on all targets.

git-svn-id: svn://svn.chromium.org/llvm-project/libcxxabi/trunk@248108 0b72dbe1-c17f-4bc7-b9db-2b4152be0356