Allow explicit pthread opt-in
When building libcxxabi in tree (i.e. in llvm/projects/libcxxabi,
along with llvm/projects/libcxx), libcxx's config_site.in doesn't
get created in a timely manner. This means that any configuration
that is normally set in libcxx's config_site.in needs to be
duplicated in libcxxabi to successfully build libcxxabi. This patch
does exactly that for the _LIBCPP_HAS_THREAD_API_PTHREAD
preprocessor define.

http://reviews.llvm.org/D20574


git-svn-id: svn://svn.chromium.org/llvm-project/libcxxabi/trunk@270732 0b72dbe1-c17f-4bc7-b9db-2b4152be0356